requirements.txt
# requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0
accelerate>=0.27.0
bitsandbytes>=0.42.0
peft>=0.8.0
trl>=0.8.0
datasets>=2.16.0
wandb>=0.16.0
tensorboard>=2.15.0
numpy>=1.24.0
tqdm>=4.65.0
Pillow>=10.0.0
opencv-python>=4.8.0
scipy>=1.10.0
scikit-learn>=1.3.0
pyyaml>=6.0
hydra-core>=1.3.0
omegaconf>=2.3.0
einops>=0.7.0
timm>=0.9.0
sentencepiece>=0.1.99
protobuf>=4.25.0
gradio>=4.0.0
requests>=2.28.0